requirements.txt
streamlit
fastapi
uvicorn
transformers
requests
Pillow
requests